Some Updates

+++ modules/reporter.py
@@ -21,6 +21,7 @@ def export_logs(app):
     settings_window = tk.Toplevel(app)
     settings_window.geometry("525x255")
     settings_window.title("Export Logs")
+    settings_window.focus_force()
 
     image_frame = tk.Frame(settings_window)
     image_frame.grid(row=0, column=1, padx=(10, 0), pady=20)
@@ -48,39 +49,50 @@ def export_logs(app):
     name_entry = ttk.Entry(settings_frame, textvariable=tk.StringVar())
     name_entry.grid(row=2, column=0, padx=(15, 0), pady=(20, 0))
 
+    def on_enter_key(event):
+        save_project(settings_window, name_entry, app)
+
+    settings_window.bind("<Return>", on_enter_key)
+
+    def on_escape_key(event):
+        settings_window.destroy()
+
+    settings_window.bind("<Escape>", on_escape_key)
+
     save_button = ttk.Button(settings_frame, text="Save", command=lambda: save_project(settings_window, name_entry, app))
     save_button.grid(row=3, column=0, pady=(35, 10))  
 
 def save_project(settings_window, name_entry, app):
-    profile_name = name_entry.get().strip()
-    if profile_name:  
-        profile_path = os.path.join("reports")
-        data_path = "data"
+    if name_entry.get():
+        profile_name = name_entry.get().strip()
+        if profile_name:  
+            profile_path = os.path.join("reports")
+            data_path = "data"
+            try:
+                os.makedirs(profile_path, exist_ok=True)  
+                for item in os.listdir(data_path):  
+                    s = os.path.join(data_path, item)
+                    d = os.path.join(profile_path, item)
+                    if os.path.isdir(s):
+                        shutil.copytree(s, d, dirs_exist_ok=True)
+                    else:
+                        shutil.copy2(s, d)
+            except:
+                pass
+
         try:
-            os.makedirs(profile_path, exist_ok=True)  
-            for item in os.listdir(data_path):  
-                s = os.path.join(data_path, item)
-                d = os.path.join(profile_path, item)
-                if os.path.isdir(s):
-                    shutil.copytree(s, d, dirs_exist_ok=True)
-                else:
-                    shutil.copy2(s, d)
+            json_file_path = os.path.join(profile_path, "sessions.json")
+            html_file_path = os.path.join(profile_path + "/" + profile_name + ".html")
+            json_to_html(json_file_path, html_file_path)
+
+            for file_name in os.listdir(profile_path):
+                if file_name.endswith('.json'):
+                    os.remove(os.path.join(profile_path, file_name))
         except:
             pass
 
-    try:
-        json_file_path = os.path.join(profile_path, "sessions.json")
-        html_file_path = os.path.join(profile_path + "/" + profile_name + ".html")
-        json_to_html(json_file_path, html_file_path)
-
-        for file_name in os.listdir(profile_path):
-            if file_name.endswith('.json'):
-                os.remove(os.path.join(profile_path, file_name))
-    except:
-        pass
-
-    settings_window.destroy()
-    app.project_saved_success()
+        settings_window.destroy()
+        app.project_saved_success()
 
 def clear_logs(app):
     profiles_path = "reports"
